# Conflicts:
#	Files/Files.csproj
#	Files/Interacts/Interaction.cs
#	Files/ViewModels/Properties/FileProperties.cs
#	Files/ViewModels/SelectedItemsPropertiesViewModel.cs

+++ Files/Filesystem/FilesystemOperations/FilesystemOperations.cs
@@ -3,7 +3,7 @@ using Files.Enums;
 using Files.Extensions;
 using Files.Filesystem.FilesystemHistory;
 using Files.Helpers;
-using Microsoft.Toolkit.Uwp.Extensions;
+using Microsoft.Toolkit.Uwp;
 using Microsoft.Toolkit.Uwp.Helpers;
 using System;
 using System.Collections.Generic;
@@ -183,7 +183,7 @@ namespace Files.Filesystem
                                 CloseButtonText = "ItemAlreadyExistsDialogCloseButtonText".GetLocalized()
                             };
 
-                            if (Interacts.Interaction.IsAnyContentDialogOpen())
+                            if (UIHelpers.IsAnyContentDialogOpen())
                             {
                                 // Only a single ContentDialog can be open at any time.
                                 return null;
@@ -249,7 +249,7 @@ namespace Files.Filesystem
                                 CloseButtonText = "ItemAlreadyExistsDialogCloseButtonText".GetLocalized()
                             };
 
-                            if (Interacts.Interaction.IsAnyContentDialogOpen())
+                            if (UIHelpers.IsAnyContentDialogOpen())
                             {
                                 // Only a single ContentDialog can be open at any time.
                                 return null;
@@ -294,8 +294,8 @@ namespace Files.Filesystem
 
                     if (copiedListedItems.Count > 0)
                     {
-                        associatedInstance.ContentPage.AddSelectedItemsOnUi(copiedListedItems);
-                        associatedInstance.ContentPage.FocusSelectedItems();
+                        associatedInstance.SlimContentPage.AddSelectedItemsOnUi(copiedListedItems);
+                        associatedInstance.SlimContentPage.FocusSelectedItems();
                     }
                 }, Windows.UI.Core.CoreDispatcherPriority.Low);
             }
@@ -411,7 +411,7 @@ namespace Files.Filesystem
                                     CloseButtonText = "ItemAlreadyExistsDialogCloseButtonText".GetLocalized()
                                 };
 
-                                if (Interacts.Interaction.IsAnyContentDialogOpen())
+                                if (UIHelpers.IsAnyContentDialogOpen())
                                 {
                                     // Only a single ContentDialog can be open at any time.
                                     return null;
@@ -474,7 +474,7 @@ namespace Files.Filesystem
                                 CloseButtonText = "ItemAlreadyExistsDialogCloseButtonText".GetLocalized()
                             };
 
-                            if (Interacts.Interaction.IsAnyContentDialogOpen())
+                            if (UIHelpers.IsAnyContentDialogOpen())
                             {
                                 // Only a single ContentDialog can be open at any time.
                                 return null;
@@ -515,8 +515,8 @@ namespace Files.Filesystem
 
                     if (movedListedItems.Count > 0)
                     {
-                        associatedInstance.ContentPage.AddSelectedItemsOnUi(movedListedItems);
-                        associatedInstance.ContentPage.FocusSelectedItems();
+                        associatedInstance.SlimContentPage.AddSelectedItemsOnUi(movedListedItems);
+                        associatedInstance.SlimContentPage.FocusSelectedItems();
                     }
                 }, Windows.UI.Core.CoreDispatcherPriority.Low);
             }
@@ -554,33 +554,64 @@ namespace Files.Filesystem
             errorCode?.Report(fsResult);
             progress?.Report(0.0f);
 
-            if (source.ItemType == FilesystemItemType.File)
+            if (permanently)
             {
-                fsResult = await associatedInstance.FilesystemViewModel.GetFileFromPathAsync(source.Path)
-                    .OnSuccess((t) => t.DeleteAsync(permanently ? StorageDeleteOption.PermanentDelete : StorageDeleteOption.Default).AsTask());
+                fsResult = (FilesystemResult)NativeFileOperationsHelper.DeleteFileFromApp(source.Path);
             }
-            else if (source.ItemType == FilesystemItemType.Directory)
+            if (!fsResult)
             {
-                fsResult = await associatedInstance.FilesystemViewModel.GetFolderFromPathAsync(source.Path)
-                    .OnSuccess((t) => t.DeleteAsync(permanently ? StorageDeleteOption.PermanentDelete : StorageDeleteOption.Default).AsTask());
+                if (source.ItemType == FilesystemItemType.File)
+                {
+                    fsResult = await associatedInstance.FilesystemViewModel.GetFileFromPathAsync(source.Path)
+                        .OnSuccess((t) => t.DeleteAsync(permanently ? StorageDeleteOption.PermanentDelete : StorageDeleteOption.Default).AsTask());
+                }
+                else if (source.ItemType == FilesystemItemType.Directory)
+                {
+                    fsResult = await associatedInstance.FilesystemViewModel.GetFolderFromPathAsync(source.Path)
+                        .OnSuccess((t) => t.DeleteAsync(permanently ? StorageDeleteOption.PermanentDelete : StorageDeleteOption.Default).AsTask());
+                }
             }
 
             errorCode?.Report(fsResult);
 
             if (fsResult == FileSystemStatusCode.Unauthorized)
             {
-                // Try again with fulltrust process
+                // Try again with fulltrust process (non admin: for shortcuts and hidden files)
                 if (associatedInstance.ServiceConnection != null)
                 {
-                    var (status, response) = await associatedInstance.ServiceConnection.SendMessageSafeAsync(new ValueSet()
-                        {
-                            { "Arguments", "FileOperation" },
-                            { "fileop", "DeleteItem" },
-                            { "filepath", source.Path },
-                            { "permanently", permanently }
-                        });
+                    var (status, response) = await associatedInstance.ServiceConnection.SendMessageForResponseAsync(new ValueSet()
+                    {
+                        { "Arguments", "FileOperation" },
+                        { "fileop", "DeleteItem" },
+                        { "filepath", source.Path },
+                        { "permanently", permanently }
+                    });
                     fsResult = (FilesystemResult)(status == AppServiceResponseStatus.Success
-                        && response.Message.Get("Success", false));
+                        && response.Get("Success", false));
+                }
+                if (!fsResult)
+                {
+                    var elevateConfirmDialog = new Files.Dialogs.ElevateConfirmDialog();
+                    var elevateConfirmResult = await elevateConfirmDialog.ShowAsync();
+                    if (elevateConfirmResult == ContentDialogResult.Primary)
+                    {
+                        if (await associatedInstance.ServiceConnection?.Elevate()) // TODO: enable this
+                        {
+                            // Try again with fulltrust process (admin)
+                            if (associatedInstance.ServiceConnection != null)
+                            {
+                                var (status, response) = await associatedInstance.ServiceConnection.SendMessageForResponseAsync(new ValueSet()
+                                {
+                                    { "Arguments", "FileOperation" },
+                                    { "fileop", "DeleteItem" },
+                                    { "filepath", source.Path },
+                                    { "permanently", permanently }
+                                });
+                                fsResult = (FilesystemResult)(status == AppServiceResponseStatus.Success
+                                    && response.Get("Success", false));
+                            }
+                        }
+                    }
                 }
             }
             else if (fsResult == FileSystemStatusCode.InUse)
@@ -723,7 +754,7 @@ namespace Files.Filesystem
                         CloseButtonText = "ItemAlreadyExistsDialogCloseButtonText".GetLocalized()
                     };
 
-                    if (Interacts.Interaction.IsAnyContentDialogOpen())
+                    if (UIHelpers.IsAnyContentDialogOpen())
                     {
                         // Only a single ContentDialog can be open at any time.
                         return null;
